Little bit of refactor

+++ 3dUtils.py
@@ -6,6 +6,7 @@ from datetime import datetime
 debug = True;
 
 datos_archivos = []
+
 def copy_folders_with_Gcode(source_dir, usb_drive):
     # Lista para almacenar los datos
     for root, dirs, files in os.walk(source_dir):
@@ -18,41 +19,55 @@ def copy_folders_with_Gcode(source_dir, usb_drive):
 
 def copy_folder(source_path, destination_path):
     if not os.path.exists(destination_path):
-        shutil.copytree(source_path, destination_path)
-        print(f"Copiado: \t{os.path.basename(source_path)}")
-        rellenarFicheroConNombres(source_path)
-        
+        copy_new_folder(source_path, destination_path)
     else:
-        removeUnused(source_path, destination_path)
-        for item in os.listdir(source_path):
-            source_item = os.path.join(source_path, item)
-            destination_item = os.path.join(destination_path, item)
+        copy_existing_folder(source_path, destination_path)
 
-            if os.path.isfile(source_item):
-                if os.path.exists(destination_item):
-                    source_mtime = os.path.getmtime(source_item)
-                    dest_mtime = os.path.getmtime(destination_item)
+def copy_new_folder(source_path, destination_path):
+    shutil.copytree(source_path, destination_path)
+    print(f"Copiado: \t{os.path.basename(source_path)}")
+    rellenar_fichero_con_nombres(source_path)
+    
+def copy_existing_folder(source_path, destination_path):
+    removeUnused(source_path, destination_path)
+    for item in os.listdir(source_path):
+        source_item = os.path.join(source_path, item)
+        destination_item = os.path.join(destination_path, item)
 
-                    if source_mtime > dest_mtime:
-                        shutil.copy2(source_item, destination_item)
-                        print(f"Sobrescrito: \t\t{os.path.basename(source_item)}")
-                        addFormatedDataToCSV(source_item)
-                else:
-                    shutil.copy2(source_item, destination_item)
-                    print(f"Copiado: \t{os.path.basename(destination_item)}")
-                    if "Done" in destination_item:
-                        parent_dir = os.path.dirname(destination_item)
-                        parent_dir = os.path.dirname(parent_dir)
-                        source_item = os.path.join(parent_dir, item)
-                        if os.path.exists(source_item):
-                            os.remove(source_item)
-                            print(f"Borrado: \t\t{os.path.basename(source_item)}")
-                    else:
-                        addFormatedDataToCSV(source_item, destination_item)
-            else:
-                copy_folder(source_item, destination_item)
+        if os.path.isfile(source_item):
+            copy_file(source_item, destination_item)
+        else:
+            copy_folder(source_item, destination_item)
 
-def rellenarFicheroConNombres(source_path):
+def copy_file(source_item, destination_item):
+    if os.path.exists(destination_item):
+        copy_existing_file(source_item, destination_item)
+    else:
+       copy_new_file(source_item, destination_item) 
+
+def copy_new_file(source_item, destination_item):
+    shutil.copy2(source_item, destination_item)
+    print(f"Copiado: \t{os.path.basename(destination_item)}")
+    if "Done" in destination_item:
+        item = os.path.dirname(destination_item)
+        parent_dir = os.path.dirname(item)
+        source_item = os.path.join(parent_dir, item)
+        if os.path.exists(source_item):
+            os.remove(source_item)
+            print(f"Borrado: \t\t{os.path.basename(source_item)}")
+    else:
+        addFormatedDataToCSV(source_item, destination_item)
+
+def copy_existing_file(source_item, destination_item):
+    source_mtime = os.path.getmtime(source_item)
+    dest_mtime = os.path.getmtime(destination_item)
+
+    if source_mtime > dest_mtime:
+        shutil.copy2(source_item, destination_item)
+        print(f"Sobrescrito: \t\t{os.path.basename(source_item)}")
+        addFormatedDataToCSV(source_item)
+
+def rellenar_fichero_con_nombres(source_path):
     for root, dirs, files in os.walk(source_path):
         for file in files:
             addFormatedDataToCSV(file, source_path)
@@ -61,8 +76,7 @@ def removeUnused(source_path, destination_path):
      for item in os.listdir(destination_path):
         source_item = os.path.join(source_path, item)
         destination_item = os.path.join(destination_path, item)
-        if os.path.exists(destination_item):
-             if os.path.isfile(destination_item):
+        if os.path.exists(destination_item) and os.path.isfile(destination_item):
                   if not os.path.exists(source_item):
                     os.remove(destination_item)
                     print(f"Borrado por obsoleto: \t{getEasyPath(destination_item)}")
@@ -73,27 +87,19 @@ def getEasyPath (path):
     return directorio_superior + "\\" + nombre_archivo 
 
 def addFormatedDataToCSV(source_item, nombrePadre):
-    proyectName = os.path.basename(os.path.dirname(nombrePadre))
-    timeToPrint = os.path.basename(source_item)[:5]
-    fileName = os.path.basename(source_item)[6:]
-    datos_archivos.append([proyectName, timeToPrint, fileName])
+    proyect_name = os.path.basename(os.path.dirname(nombrePadre))
+    time_to_print = os.path.basename(source_item)[:5]
+    file_name = os.path.basename(source_item)[6:]
+    datos_archivos.append([proyect_name, time_to_print, file_name])
 
 def writeXlsx():
-
     # Abre el archivo Excel
     workbook = openpyxl.load_workbook('ejemplo.xlsx')
 
     # Selecciona la hoja en la que deseas agregar datos
     sheet = workbook['Resultado']  # O puedes seleccionar una hoja específica por nombre
     
-    # Encuentra la primera fila vacía en la columna A
-    columna_A = sheet['A']
-    primera_fila_vacia = None
-
-    for fila, celda in enumerate(columna_A, start=1):
-        if not celda.value:
-            primera_fila_vacia = fila
-            break
+    primera_fila_vacia = get_place_to_write(sheet)
 
     # Agrega los nuevos datos al final de la hoja
     numArchivos = 0
@@ -118,8 +124,14 @@ def writeXlsx():
     # Cierra el archivo Excel
     workbook.close()
 
-    print("Datos escritos al final del archivo tu_archivo.xlsx")
+def get_place_to_write(sheet):
+    # Encuentra la primera fila vacía en la columna A
+    columna_A = sheet['A']
+    primera_fila_vacia = None
 
+    for fila, celda in enumerate(columna_A, start=1):
+        if not celda.value:
+            return fila
 
 if __name__ == "__main__":
     source_directory = "C:\\Users\\yager\\Documents\\Impresion3D"  # Cambia esta ruta a tu directorio de origen
@@ -136,4 +148,3 @@ if __name__ == "__main__":
             input("PROGRAMA FINALIZADO")
         else: 
             input("FALTA USB")
-
